requirements.txt
google-genai
pandas
openpyxl
python-dotenv
pydantic
torch
torchvision
transformers
datasets
scikit-learn
tqdm